pat stacks
