+++ src/pat/stacks/index.html
@@ -1,39 +1,48 @@
 <!DOCTYPE html>
 <html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <title>Demo page</title>
-    <link rel="stylesheet" href="/style/common.css" type="text/css">
-    <link rel="stylesheet" href="/style/patterns.css" type="text/css">
-        <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
+    <link rel="stylesheet" href="/style/common.css" type="text/css" />
+    <link rel="stylesheet" href="/style/patterns.css" type="text/css" />
+    <script src="/bundle-vendors.js" type="text/javascript" charset="utf-8"></script>
+    <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
     <style>
-        .navigation.tabs {
-            background-color: white;
-        }
+      .navigation.tabs {
+        background-color: white;
+      }
     </style>
-</head>
-<body>
-<h3>The stacks pattern can be used to make tabs</h3>
-<nav class="tabs">
-    <a href="#sheet-1" class="current">Tab one</a> <a href="#sheet-2">Tab two</a> <a href="#sheet-3">Tab three</a>
-</nav>
-<article class="pat-stacks">
-    <section id="sheet-1">
+  </head>
+  <body>
+    <h3>The stacks pattern can be used to make tabs</h3>
+    <nav class="tabs">
+      <a href="#sheet-1" class="current">Tab one</a>
+      <a href="#sheet-2">Tab two</a> <a href="#sheet-3">Tab three</a>
+    </nav>
+    <article class="pat-stacks">
+      <section id="sheet-1">
         <h4>Sheet one</h4>
         <p>
-            Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo.
+          Sed ut perspiciatis unde omnis iste natus error sit voluptatem
+          accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae
+          ab illo inventore veritatis et quasi architecto beatae vitae dicta
+          sunt explicabo.
         </p>
-    </section>
-    <section id="sheet-2">
+      </section>
+      <section id="sheet-2">
         <h4>Sheet two</h4>
         <p>
-            Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt.
+          Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut
+          fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem
+          sequi nesciunt.
         </p>
-    </section>
-    <section id="sheet-3">
+      </section>
+      <section id="sheet-3">
         <h4>Sheet three</h4>
         <p>
-            Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem.
+          Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet,
+          consectetur, adipisci velit, sed quia non numquam eius modi tempora
+          incidunt ut labore et dolore magnam aliquam quaerat voluptatem.
         </p>
     </section>
 </article>
@@ -96,7 +105,7 @@ pattern lets the tabs stick to the top of the page when you scroll down.
                 our dormitory last night. Her attitude toward the captive
                 was most harsh and brutal. When she held her, she sunk her
                 rudimentary nails into the poor girl's flesh, or twisted
-                her arm in a most painful manner. 
+                her arm in a most painful manner.
                 </p>
             </article>
         </div>
